+++ VulkanGameEngineLevelEditorInterlopDLL/VulkanSystemDLL.h
@@ -50,14 +50,14 @@ extern "C" {
 	DLL_EXPORT GraphicsSystem			   VulkanSystem_RendererSetUp(void* windowHandle, VkInstance& instance, VkSurfaceKHR& surface);
 	DLL_EXPORT VkCommandBuffer			   VulkanSystem_BeginSingleUseCommand();
 	DLL_EXPORT void						   VulkanSystem_EndSingleUseCommand(VkCommandBuffer commandBuffer);
-	DLL_EXPORT VkPresentModeKHR*		   VulkanSystem_GetSurfacePresentModes(VkPhysicalDevice physicalDevice, VkSurfaceKHR surface, uint32* outCount);
+	DLL_EXPORT VkPresentModeKHR*		   VulkanSystem_GetSurfacePresentModes(VkPhysicalDevice physicalDevice, VkSurfaceKHR surface, size_t* outCount);
 	DLL_EXPORT VkSurfaceCapabilitiesKHR    VulkanSystem_GetSurfaceCapabilities(VkPhysicalDevice physicalDevice, VkSurfaceKHR surface);
 	DLL_EXPORT VkPhysicalDeviceProperties  VulkanSystem_GetPhysicalDeviceProperties(VkPhysicalDevice physicalDevice);
 	DLL_EXPORT VkPhysicalDeviceFeatures    VulkanSystem_GetPhysicalDeviceFeatures(VkPhysicalDevice physicalDevice);
 	DLL_EXPORT VkPhysicalDeviceFeatures2   VulkanSystem_GetPhysicalDeviceFeatures2(VkPhysicalDevice physicalDevice);
-	DLL_EXPORT VkPhysicalDevice*		   VulkanSystem_GetPhysicalDeviceList(VkInstance instance, uint32* outCount);
-	DLL_EXPORT VkSurfaceFormatKHR*		   VulkanSystem_GetPhysicalDeviceFormats(VkPhysicalDevice physicalDevice, VkSurfaceKHR surface, uint32* outCount);
-	DLL_EXPORT VkPresentModeKHR*		   VulkanSystem_GetPhysicalDevicePresentModes(VkPhysicalDevice physicalDevice, VkSurfaceKHR surface, uint32* outCount);
+	DLL_EXPORT VkPhysicalDevice*		   VulkanSystem_GetPhysicalDeviceList(VkInstance instance, size_t* outCount);
+	DLL_EXPORT VkSurfaceFormatKHR*		   VulkanSystem_GetPhysicalDeviceFormats(VkPhysicalDevice physicalDevice, VkSurfaceKHR surface, size_t* outCount);
+	DLL_EXPORT VkPresentModeKHR*		   VulkanSystem_GetPhysicalDevicePresentModes(VkPhysicalDevice physicalDevice, VkSurfaceKHR surface, size_t* outCount);
 	DLL_EXPORT void						   VulkanSystem_DestroyRenderer();
 #ifdef __cplusplus
 }
